Refactor ApptainerFacade to use Backend enum and split Lima logic into separate module
