+++ crates/containers-internal/src/apptainer.rs
@@ -1,4 +1,5 @@
 pub mod facade;
+pub(crate) mod lima;
 
 #[cfg(test)]
 mod tests;
